update main.py in replit
