+++ .replit
@@ -1,0 +1,4 @@
+modules = ["nodejs-20", "python-3.12", "web"]
+
+[nix]
+channel = "stable-24_05"
